requirements.txt
asgiref==3.2.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
DateTime==4.3
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.0.6
django-allauth==0.41.0
django-cors-headers==2.2.0
django-heroku==0.3.1
django-phone-field==1.8.0
django-rest-auth==0.9.5
django-rest-authtoken==1.2.4
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.9
names==0.3.0
nose==1.3.7
oauthlib==3.1.0
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2018.4
Random-Word==1.0.4
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
whitenoise==5.0.1
zope.interface==5.0.1
